refactor(replay): split replayService.ts into per-path modules
1616 lines carrying four responsibilities becomes: a 184-line
ReplayService entry surface delegating to replayFoldPath.ts,
replayMapPath.ts, and replayOptimizedPath.ts (module-level functions
over an explicit ReplayContext), with replayDiscovery.ts, replayLog.ts,
and pMapLimited.ts extracted for the shared pieces.
clearFailedBatchMarkers moved into replayMarkers.ts. No behaviour
change; no re-exports; public ReplayService import path unchanged.
