+++ langwatch/src/server/event-sourcing/replay/replayService.ts
@@ -2,61 +2,26 @@ import type { ClickHouseClient } from "@clickhouse/client";
 import type IORedis from "ioredis";
 import type {
   RegisteredFoldProjection,
-  RegisteredMapProjection,
-  ReplayProgress,
   ReplayConfig,
   ReplayCallbacks,
   ReplayResult,
   DiscoveryResult,
-  BatchPhase,
-  BatchCompleteInfo,
-  ProjectionKind,
+  ReplayContext,
 } from "./types";
-import type { CutoffInfo, DiscoveredAggregate, ReplayEvent } from "./replayEventLoader";
-import { isAtOrBeforeCutoff } from "./replayConstants";
-import type { OccurredAtBounds } from "./replayEventLoader";
-import {
-  discoverAffectedAggregates,
-  countEventsForAggregates,
-  batchLoadAggregateEvents,
-  getBoundedCutoffs,
-  loadEventsForAggregatesBulk,
-} from "./replayEventLoader";
-import {
-  aggregateKey,
-  markPendingBatch,
-  markCutoffBatch,
-  markCompletedBatch,
-  unmarkBatch,
-  removeInFlightMarkers,
-  getCompletedSet,
-  getCutoffMarkers,
-  removeStaleMarker,
-  cleanupAll,
-  hasPreviousRun,
-} from "./replayMarkers";
-import { pauseProjection, unpauseProjection, waitForActiveJobs, waitForAllActiveJobs } from "./replayDrain";
-import { FoldAccumulator, MapAccumulator } from "./replayExecutor";
+import { cleanupAll, hasPreviousRun } from "./replayMarkers";
+import type { ReplayLogWriter } from "./replayLog";
+import { nullLog } from "./replayLog";
+import { discoverProjectionAggregates } from "./replayDiscovery";
+import { replayFoldProjection } from "./replayFoldPath";
+import { replayMapProjection } from "./replayMapPath";
+import { replayOptimized } from "./replayOptimizedPath";
 import type { RetentionPolicyResolver } from "../../data-retention/retentionPolicyResolver";
 
-/** Minimal log interface — CLI provides concrete implementation. */
-export interface ReplayLogWriter {
-  write(entry: Record<string, unknown>): void;
-}
-
-/** No-op log for when no logging is needed. */
-const nullLog: ReplayLogWriter = { write() {} };
-
-/**
- * Emit replay-phase progress once per this many completed aggregates (plus
- * once at the end of the batch). Every emit fans out to the progress callback
- * — which the ops layer persists to Redis in multiple round trips — so
- * per-aggregate emits (1000/batch) hammered Redis for no operator benefit.
- */
-const PROGRESS_EMIT_EVERY_AGGREGATES = 100;
-
 export class ReplayService {
-  constructor(private readonly deps: {
+  /** Shared dependencies handed to the path implementations. */
+  private readonly ctx: ReplayContext;
+
+  constructor(deps: {
     clickhouseClientResolver: (tenantId: string) => Promise<ClickHouseClient>;
     redis: IORedis;
     /**
@@ -66,72 +31,13 @@ export class ReplayService {
      * behaviour (and the NullReplayRepository test path).
      */
     retentionPolicyResolver?: RetentionPolicyResolver;
-  }) {}
-
-  /** Accumulator options carrying the retention resolver (if wired). */
-  private get accumulatorOpts(): { retentionResolver?: RetentionPolicyResolver } {
-    return { retentionResolver: this.deps.retentionPolicyResolver };
-  }
-
-  /**
-   * Restrict discovered aggregates to a caller-supplied `aggregateIds` allow-list
-   * (single-/scoped-aggregate replay). Mutates `byTenant` in place to stay in
-   * sync and returns the filtered `allAggregates`. A no-op when the list is
-   * empty/absent (full replay). The optimized path applies the same filter
-   * against its aggregate→projection map.
-   */
-  private static filterDiscoveredByAggregateIds(
-    allAggregates: DiscoveredAggregate[],
-    byTenant: Map<string, DiscoveredAggregate[]>,
-    aggregateIds?: string[],
-  ): DiscoveredAggregate[] {
-    if (!aggregateIds || aggregateIds.length === 0) return allAggregates;
-    const allowed = new Set(aggregateIds);
-    for (const [tid, aggs] of byTenant) {
-      const kept = aggs.filter((a) => allowed.has(a.aggregateId));
-      if (kept.length > 0) byTenant.set(tid, kept);
-      else byTenant.delete(tid);
-    }
-    return allAggregates.filter((a) => allowed.has(a.aggregateId));
-  }
-
-  private async resolveClient(tenantId?: string): Promise<ClickHouseClient> {
-    return this.deps.clickhouseClientResolver(tenantId ?? "default");
-  }
-
-  /**
-   * Best-effort marker cleanup for a batch that errored (or was abandoned by
-   * cancellation) mid-flight. Its aggregates were never replayed, so their
-   * pending/cutoff markers are removed — returning them to unconditional live
-   * processing right away instead of deferring their live events until the
-   * 7-day marker TTL lapses. Done markers and completed-set entries from
-   * previously completed batches are left intact so an operator re-run still
-   * skips completed aggregates. Never throws: a marker-cleanup failure must
-   * not mask the original batch error.
-   */
-  private async clearFailedBatchMarkers({
-    projectionNames,
-    aggKeys,
-    log,
-  }: {
-    projectionNames: string[];
-    aggKeys: string[];
-    log: ReplayLogWriter;
-  }): Promise<void> {
-    try {
-      await removeInFlightMarkers({
-        redis: this.deps.redis,
-        projectionNames,
-        aggKeys,
-      });
-    } catch (cleanupError) {
-      log.write({
-        step: "error",
-        error: `failed to clear replay markers for failed batch: ${
-          cleanupError instanceof Error ? cleanupError.message : String(cleanupError)
-        }`,
-      });
-    }
+  }) {
+    this.ctx = {
+      redis: deps.redis,
+      resolveClient: (tenantId?: string) =>
+        deps.clickhouseClientResolver(tenantId ?? "default"),
+      accumulatorOpts: { retentionResolver: deps.retentionPolicyResolver },
+    };
   }
 
   async discover({
@@ -143,31 +49,12 @@ export class ReplayService {
     since: string;
     tenantId?: string;
   }): Promise<DiscoveryResult> {
-    const client = await this.resolveClient(tenantId);
-    const sinceMs = new Date(since).getTime();
-    const [aggregates, totalEvents] = await Promise.all([
-      discoverAffectedAggregates({
-        client,
-        eventTypes: projection.definition.eventTypes,
-        sinceMs,
-        tenantId,
-      }),
-      countEventsForAggregates({
-        client,
-        eventTypes: projection.definition.eventTypes,
-        sinceMs,
-        tenantId,
-      }),
-    ]);
-
-    const byTenant = new Map<string, DiscoveredAggregate[]>();
-    for (const agg of aggregates) {
-      const list = byTenant.get(agg.tenantId) ?? [];
-      list.push(agg);
-      byTenant.set(agg.tenantId, list);
-    }
-
-    return { aggregates, byTenant, tenantCount: byTenant.size, totalEvents };
+    return discoverProjectionAggregates({
+      resolveClient: this.ctx.resolveClient,
+      projection,
+      since,
+      tenantId,
+    });
   }
 
   async replay(
@@ -189,7 +76,8 @@ export class ReplayService {
 
     for (let pi = 0; pi < config.projections.length; pi++) {
       const projection = config.projections[pi]!;
-      const result = await this.replayProjection({
+      const result = await replayFoldProjection({
+        ctx: this.ctx,
         projection,
         projectionIndex: pi,
         totalProjections,
@@ -216,7 +104,8 @@ export class ReplayService {
     if (totalBatchErrors === 0) {
       for (let mi = 0; mi < mapProjections.length; mi++) {
         const projection = mapProjections[mi]!;
-        const result = await this.replayMapProjection({
+        const result = await replayMapProjection({
+          ctx: this.ctx,
           projection,
           projectionIndex: config.projections.length + mi,
           totalProjections,
@@ -259,7 +148,7 @@ export class ReplayService {
 
       for (const tenantId of tenantTargets) {
         try {
-          const client = await this.resolveClient(tenantId);
+          const client = await this.ctx.resolveClient(tenantId);
           for (const table of tables) {
             await client.command({ query: `OPTIMIZE TABLE ${table}` });
             callbacks?.log?.write({ step: "optimize", table, tenant: tenantId });
@@ -278,1339 +167,18 @@ export class ReplayService {
     };
   }
 
-  private async replayProjection({
-    projection,
-    projectionIndex,
-    totalProjections,
-    tenantIds,
-    aggregateIds,
-    since,
-    batchSize,
-    aggregateBatchSize,
-    dryRun,
-    log,
-    onProgress,
-    onBatchComplete,
-  }: {
-    projection: RegisteredFoldProjection;
-    projectionIndex: number;
-    totalProjections: number;
-    tenantIds: string[];
-    aggregateIds?: string[];
-    since: string;
-    batchSize: number;
-    aggregateBatchSize: number;
-    dryRun: boolean;
-    log: ReplayLogWriter;
-    onProgress?: (progress: ReplayProgress) => void;
-    onBatchComplete?: (info: BatchCompleteInfo) => void;
-  }): Promise<ReplayResult & { touchedTenants: string[] }> {
-    const redis = this.deps.redis;
-    const startTime = Date.now();
-
-    // Discover aggregates — when tenantIds is empty, discover across ALL tenants
-    let allAggregates: DiscoveredAggregate[] = [];
-    const byTenant = new Map<string, DiscoveredAggregate[]>();
-
-    const discoveryTargets = tenantIds.length > 0 ? tenantIds : [undefined];
-    for (const tenantId of discoveryTargets) {
-      const discovery = await this.discover({ projection, since, tenantId });
-      allAggregates = allAggregates.concat(discovery.aggregates);
-      for (const [tid, aggs] of discovery.byTenant) {
-        const existing = byTenant.get(tid) ?? [];
-        byTenant.set(tid, existing.concat(aggs));
-      }
-    }
-
-    // Scoped replay: keep only the requested aggregates (no-op for full replay).
-    allAggregates = ReplayService.filterDiscoveredByAggregateIds(
-      allAggregates,
-      byTenant,
-      aggregateIds,
-    );
-
-    if (allAggregates.length === 0) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0, touchedTenants: [] };
-    }
-
-    if (dryRun) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0, touchedTenants: [] };
-    }
-
-    // Get completed set for resume support
-    const completedSet = await getCompletedSet({ redis, projectionName: projection.projectionName });
-
-    // Remove stale markers
-    const staleMarkers = await getCutoffMarkers({ redis, projectionName: projection.projectionName });
-    for (const aggKey of staleMarkers.keys()) {
-      await removeStaleMarker({ redis, projectionName: projection.projectionName, aggKey });
-    }
-
-    let aggregatesCompleted = 0;
-    let totalEventsReplayed = 0;
-    let skippedCount = 0;
-    let batchErrors = 0;
-    let firstError: string | undefined;
-
-    const tenants = [...byTenant.entries()];
-
-    for (const [tenantId, tenantAggregates] of tenants) {
-      const client = await this.resolveClient(tenantId);
-      const remaining: DiscoveredAggregate[] = [];
-      for (const agg of tenantAggregates) {
-        if (completedSet.has(aggregateKey(agg))) {
-          skippedCount++;
-          aggregatesCompleted++;
-        } else {
-          remaining.push(agg);
-        }
-      }
-
-      const totalBatches = Math.ceil(remaining.length / aggregateBatchSize);
-
-      for (let i = 0; i < remaining.length; i += aggregateBatchSize) {
-        const batch = remaining.slice(i, i + aggregateBatchSize);
-        const batchNum = Math.floor(i / aggregateBatchSize) + 1;
-        const batchStartTime = Date.now();
-
-        const progress: ReplayProgress = {
-          phase: "replaying",
-          currentProjectionName: projection.projectionName,
-          currentProjectionKind: "fold",
-          currentProjectionIndex: projectionIndex,
-          totalProjections,
-          totalAggregates: allAggregates.length,
-          tenantCount: byTenant.size,
-          currentBatch: batchNum,
-          totalBatches,
-          batchAggregates: batch.length,
-          batchPhase: "mark",
-          batchEventsProcessed: 0,
-          aggregatesCompleted,
-          totalEventsReplayed,
-          elapsedSec: (Date.now() - startTime) / 1000,
-          skippedCount,
-          batchErrors,
-          firstError,
-        };
-
-        const emit = () => {
-          progress.elapsedSec = (Date.now() - startTime) / 1000;
-          onProgress?.({ ...progress });
-        };
-
-        emit();
-
-        try {
-          const result = await this.replayBatch({
-            client,
-            redis,
-            projection,
-            batch,
-            tenantId,
-            batchSize,
-            log,
-            onBatchPhase: (phase, eventsProcessed) => {
-              progress.batchPhase = phase;
-              if (eventsProcessed !== undefined) {
-                progress.batchEventsProcessed = eventsProcessed;
-                progress.totalEventsReplayed = totalEventsReplayed + eventsProcessed;
-              }
-              emit();
-            },
-          });
-
-          totalEventsReplayed += result.eventsReplayed;
-          aggregatesCompleted += batch.length;
-
-          onBatchComplete?.({
-            projectionName: projection.projectionName,
-            projectionKind: "fold",
-            batchNum,
-            totalBatches,
-            aggregatesInBatch: batch.length,
-            eventsInBatch: result.eventsReplayed,
-            durationSec: (Date.now() - batchStartTime) / 1000,
-          });
-        } catch (error) {
-          batchErrors++;
-          const errorMsg = error instanceof Error ? error.message : String(error);
-          if (!firstError) firstError = errorMsg;
-          log.write({
-            step: "error",
-            tenant: tenantId,
-            aggregate: `batch ${batchNum}`,
-            error: errorMsg,
-          });
-
-          await this.clearFailedBatchMarkers({
-            projectionNames: [projection.projectionName],
-            aggKeys: batch.map((agg) => aggregateKey(agg)),
-            log,
-          });
-
-          progress.batchErrors = batchErrors;
-          progress.firstError = firstError;
-          emit();
-
-          await unpauseProjection({
-            redis,
-            pauseKey: projection.pauseKey,
-          }).catch(() => {});
-
-          return {
-            aggregatesReplayed: aggregatesCompleted - skippedCount,
-            totalEvents: totalEventsReplayed,
-            batchErrors,
-            firstError,
-            touchedTenants: tenants.map(([tid]) => tid),
-          };
-        }
-      }
-    }
-
-    await cleanupAll({ redis, projectionName: projection.projectionName });
-
-    return {
-      aggregatesReplayed: aggregatesCompleted - skippedCount,
-      totalEvents: totalEventsReplayed,
-      batchErrors,
-      firstError,
-      touchedTenants: tenants.map(([tid]) => tid),
-    };
-  }
-
-  /**
-   * Replays a single batch of aggregates through the 7-phase cycle.
-   *
-   * Events are streamed page-by-page through a FoldAccumulator — only fold
-   * states (bounded by batch.length) stay in memory, not the raw events.
-   */
-  private async replayBatch({
-    client,
-    redis,
-    projection,
-    batch,
-    tenantId,
-    batchSize,
-    log,
-    onBatchPhase,
-  }: {
-    client: ClickHouseClient;
-    redis: IORedis;
-    projection: RegisteredFoldProjection;
-    batch: DiscoveredAggregate[];
-    tenantId: string;
-    batchSize: number;
-    log: ReplayLogWriter;
-    onBatchPhase: (phase: BatchPhase, eventsProcessed?: number) => void;
-  }): Promise<{ eventsReplayed: number }> {
-    const projectionName = projection.projectionName;
-    const aggKeys = batch.map((agg) => aggregateKey(agg));
-
-    // 1. MARK
-    onBatchPhase("mark");
-    await markPendingBatch({ redis, projectionName, aggKeys });
-    log.write({ step: "mark-batch", tenant: tenantId, count: batch.length });
-
-    // 2. PAUSE
-    onBatchPhase("pause");
-    await pauseProjection({ redis, pauseKey: projection.pauseKey });
-
-    // 3. DRAIN
-    onBatchPhase("drain");
-    const drainStart = Date.now();
-    await waitForActiveJobs({ redis, aggregates: batch, projectionName, kind: "fold" });
-    log.write({ step: "drain-batch", tenant: tenantId, count: batch.length, durationMs: Date.now() - drainStart });
-
-    // 4. CUTOFF — occurred-at bounds first (over all events of the batch's
-    //    aggregates) so the cutoff + load queries prune event_log's weekly
-    //    partitions instead of scanning cold storage. See
-    //    getAggregateOccurredAtBounds for why this bound is safe, and
-    //    getBoundedCutoffs for the zero-event short-circuit.
-    onBatchPhase("cutoff");
-    const { cutoffs, occurredAtBounds } = await getBoundedCutoffs({
-      client,
-      tenantId,
-      aggregateTypes: [...new Set(batch.map((a) => a.aggregateType))],
-      aggregateIds: batch.map((a) => a.aggregateId),
-      eventTypes: projection.definition.eventTypes,
-    });
-
-    const withCutoffKeys: string[] = [];
-    const withoutCutoffKeys: string[] = [];
-    for (const aggKey of aggKeys) {
-      if (cutoffs.has(aggKey)) {
-        withCutoffKeys.push(aggKey);
-      } else {
-        withoutCutoffKeys.push(aggKey);
-      }
-    }
-
-    log.write({ step: "cutoff-batch", tenant: tenantId, count: batch.length, withEvents: withCutoffKeys.length });
-
-    if (withoutCutoffKeys.length > 0) {
-      await unmarkBatch({ redis, projectionName, aggKeys: withoutCutoffKeys });
-    }
-
-    if (withCutoffKeys.length === 0) {
-      onBatchPhase("unmark");
-      await unpauseProjection({ redis, pauseKey: projection.pauseKey });
-      return { eventsReplayed: 0 };
-    }
-
-    await markCutoffBatch({ redis, projectionName, cutoffs });
-
-    // 5. REPLAY — stream events page-by-page through the fold accumulator.
-    //    Only fold states (bounded by batch size) stay in memory.
-    onBatchPhase("replay", 0);
-    const accumulator = new FoldAccumulator(projection.definition, this.accumulatorOpts);
-    const maxCutoffEventId = [...cutoffs.values()]
-      .map((c) => c.eventId)
-      .reduce((a, b) => (a > b ? a : b));
-    const aggregateIds = batch
-      .filter((a) => cutoffs.has(aggregateKey(a)))
-      .map((a) => a.aggregateId);
-
-    let cursorEventId = "";
-    const replayStart = Date.now();
-
-    while (true) {
-      const events = await batchLoadAggregateEvents({
-        client,
-        tenantId,
-        aggregateIds,
-        eventTypes: projection.definition.eventTypes,
-        maxCutoffEventId,
-        cursorEventId,
-        batchSize,
-        occurredAtBounds,
-      });
-
-      if (events.length === 0) break;
-
-      for (const e of events) {
-        const key = aggregateKey({
-          tenantId: e.tenantId,
-          aggregateType: e.aggregateType,
-          aggregateId: e.aggregateId,
-        });
-        const cutoff = cutoffs.get(key);
-        if (cutoff != null && isAtOrBeforeCutoff(e.timestamp, e.id, cutoff.timestamp, cutoff.eventId)) {
-          accumulator.apply(e);
-          onBatchPhase("replay", accumulator.processed);
-        }
-      }
-
-      const lastEvent = events[events.length - 1];
-      if (lastEvent) cursorEventId = lastEvent.id;
-      if (events.length < batchSize) break;
-    }
-
-    // 6. WRITE — flush accumulated fold states to ClickHouse
-    const totalBatchEvents = accumulator.processed;
-    onBatchPhase("write", totalBatchEvents);
-    await accumulator.flush();
-
-    log.write({
-      step: "replay-batch",
-      tenant: tenantId,
-      count: withCutoffKeys.length,
-      eventsProcessed: totalBatchEvents,
-      durationMs: Date.now() - replayStart,
-    });
-
-    // 7. COMPLETE + UNPAUSE — replace each replayed aggregate's active cutoff
-    //    marker with a terminal `done:` marker (rather than deleting it) so a
-    //    job staged but never active during the pause is still skipped for
-    //    events at/before the cutoff after unpause, instead of double-writing.
-    onBatchPhase("unmark", totalBatchEvents);
-    await markCompletedBatch({ redis, projectionName, cutoffs });
-    await unpauseProjection({ redis, pauseKey: projection.pauseKey });
-    log.write({ step: "unmark-batch", tenant: tenantId, count: withCutoffKeys.length });
-
-    return { eventsReplayed: totalBatchEvents };
-  }
-
-  /**
-   * Replays a single map projection across discovered aggregates.
-   *
-   * Mirrors `replayProjection` but:
-   * - No per-aggregate accumulator — each event is transformed and appended
-   *   immediately (`projection.map(event) → store.append(record, ctx)`).
-   * - "Replay" and "Write" phases collapse into the per-event loop.
-   */
-  private async replayMapProjection({
-    projection,
-    projectionIndex,
-    totalProjections,
-    tenantIds,
-    aggregateIds,
-    since,
-    batchSize,
-    aggregateBatchSize,
-    dryRun,
-    log,
-    onProgress,
-    onBatchComplete,
-  }: {
-    projection: RegisteredMapProjection;
-    projectionIndex: number;
-    totalProjections: number;
-    tenantIds: string[];
-    aggregateIds?: string[];
-    since: string;
-    batchSize: number;
-    aggregateBatchSize: number;
-    dryRun: boolean;
-    log: ReplayLogWriter;
-    onProgress?: (progress: ReplayProgress) => void;
-    onBatchComplete?: (info: BatchCompleteInfo) => void;
-  }): Promise<ReplayResult & { touchedTenants: string[] }> {
-    const redis = this.deps.redis;
-    const startTime = Date.now();
-
-    // Discover aggregates — reuses the same event_log scan as folds.
-    let allAggregates: DiscoveredAggregate[] = [];
-    const byTenant = new Map<string, DiscoveredAggregate[]>();
-
-    const discoveryTargets = tenantIds.length > 0 ? tenantIds : [undefined];
-    for (const tenantId of discoveryTargets) {
-      const discovery = await this.discover({
-        // Cast: discover() is typed for RegisteredFoldProjection but only reads
-        // `definition.eventTypes`, which map projections also expose.
-        projection: projection as unknown as RegisteredFoldProjection,
-        since,
-        tenantId,
-      });
-      allAggregates = allAggregates.concat(discovery.aggregates);
-      for (const [tid, aggs] of discovery.byTenant) {
-        const existing = byTenant.get(tid) ?? [];
-        byTenant.set(tid, existing.concat(aggs));
-      }
-    }
-
-    // Scoped replay: keep only the requested aggregates (no-op for full replay).
-    allAggregates = ReplayService.filterDiscoveredByAggregateIds(
-      allAggregates,
-      byTenant,
-      aggregateIds,
-    );
-
-    if (allAggregates.length === 0 || dryRun) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0, touchedTenants: [] };
-    }
-
-    const completedSet = await getCompletedSet({ redis, projectionName: projection.projectionName });
-
-    const staleMarkers = await getCutoffMarkers({ redis, projectionName: projection.projectionName });
-    for (const aggKey of staleMarkers.keys()) {
-      await removeStaleMarker({ redis, projectionName: projection.projectionName, aggKey });
-    }
-
-    let aggregatesCompleted = 0;
-    let totalEventsReplayed = 0;
-    let skippedCount = 0;
-    let batchErrors = 0;
-    let firstError: string | undefined;
-
-    const tenants = [...byTenant.entries()];
-
-    for (const [tenantId, tenantAggregates] of tenants) {
-      const client = await this.resolveClient(tenantId);
-      const remaining: DiscoveredAggregate[] = [];
-      for (const agg of tenantAggregates) {
-        if (completedSet.has(aggregateKey(agg))) {
-          skippedCount++;
-          aggregatesCompleted++;
-        } else {
-          remaining.push(agg);
-        }
-      }
-
-      const totalBatches = Math.ceil(remaining.length / aggregateBatchSize);
-
-      for (let i = 0; i < remaining.length; i += aggregateBatchSize) {
-        const batch = remaining.slice(i, i + aggregateBatchSize);
-        const batchNum = Math.floor(i / aggregateBatchSize) + 1;
-        const batchStartTime = Date.now();
-
-        const progress: ReplayProgress = {
-          phase: "replaying",
-          currentProjectionName: projection.projectionName,
-          currentProjectionKind: "map",
-          currentProjectionIndex: projectionIndex,
-          totalProjections,
-          totalAggregates: allAggregates.length,
-          tenantCount: byTenant.size,
-          currentBatch: batchNum,
-          totalBatches,
-          batchAggregates: batch.length,
-          batchPhase: "mark",
-          batchEventsProcessed: 0,
-          aggregatesCompleted,
-          totalEventsReplayed,
-          elapsedSec: (Date.now() - startTime) / 1000,
-          skippedCount,
-          batchErrors,
-          firstError,
-        };
-
-        const emit = () => {
-          progress.elapsedSec = (Date.now() - startTime) / 1000;
-          onProgress?.({ ...progress });
-        };
-
-        emit();
-
-        try {
-          const result = await this.replayMapBatch({
-            client,
-            redis,
-            projection,
-            batch,
-            tenantId,
-            batchSize,
-            log,
-            onBatchPhase: (phase, eventsProcessed) => {
-              progress.batchPhase = phase;
-              if (eventsProcessed !== undefined) {
-                progress.batchEventsProcessed = eventsProcessed;
-                progress.totalEventsReplayed = totalEventsReplayed + eventsProcessed;
-              }
-              emit();
-            },
-          });
-
-          totalEventsReplayed += result.eventsReplayed;
-          aggregatesCompleted += batch.length;
-
-          onBatchComplete?.({
-            projectionName: projection.projectionName,
-            projectionKind: "map",
-            batchNum,
-            totalBatches,
-            aggregatesInBatch: batch.length,
-            eventsInBatch: result.eventsReplayed,
-            durationSec: (Date.now() - batchStartTime) / 1000,
-          });
-        } catch (error) {
-          batchErrors++;
-          const errorMsg = error instanceof Error ? error.message : String(error);
-          if (!firstError) firstError = errorMsg;
-          log.write({
-            step: "error",
-            tenant: tenantId,
-            aggregate: `map-batch ${batchNum}`,
-            error: errorMsg,
-          });
-
-          await this.clearFailedBatchMarkers({
-            projectionNames: [projection.projectionName],
-            aggKeys: batch.map((agg) => aggregateKey(agg)),
-            log,
-          });
-
-          progress.batchErrors = batchErrors;
-          progress.firstError = firstError;
-          emit();
-
-          await unpauseProjection({
-            redis,
-            pauseKey: projection.pauseKey,
-          }).catch(() => {});
-
-          return {
-            aggregatesReplayed: aggregatesCompleted - skippedCount,
-            totalEvents: totalEventsReplayed,
-            batchErrors,
-            firstError,
-            touchedTenants: tenants.map(([tid]) => tid),
-          };
-        }
-      }
-    }
-
-    await cleanupAll({ redis, projectionName: projection.projectionName });
-
-    return {
-      aggregatesReplayed: aggregatesCompleted - skippedCount,
-      totalEvents: totalEventsReplayed,
-      batchErrors,
-      firstError,
-      touchedTenants: tenants.map(([tid]) => tid),
-    };
-  }
-
-  /**
-   * Replays a single batch of aggregates through a map projection.
-   *
-   * Same 7-phase cycle as `replayBatch`, but the REPLAY phase applies
-   * `projection.map(event)` and appends to `projection.store` per event,
-   * without accumulating state. There's no WRITE step — appends happen inline.
-   */
-  private async replayMapBatch({
-    client,
-    redis,
-    projection,
-    batch,
-    tenantId,
-    batchSize,
-    log,
-    onBatchPhase,
-  }: {
-    client: ClickHouseClient;
-    redis: IORedis;
-    projection: RegisteredMapProjection;
-    batch: DiscoveredAggregate[];
-    tenantId: string;
-    batchSize: number;
-    log: ReplayLogWriter;
-    onBatchPhase: (phase: BatchPhase, eventsProcessed?: number) => void;
-  }): Promise<{ eventsReplayed: number }> {
-    const projectionName = projection.projectionName;
-    const aggKeys = batch.map((agg) => aggregateKey(agg));
-
-    // 1. MARK
-    onBatchPhase("mark");
-    await markPendingBatch({ redis, projectionName, aggKeys });
-    log.write({ step: "mark-batch", tenant: tenantId, count: batch.length, kind: "map" });
-
-    // 2. PAUSE
-    onBatchPhase("pause");
-    await pauseProjection({ redis, pauseKey: projection.pauseKey });
-
-    // 3. DRAIN
-    onBatchPhase("drain");
-    const drainStart = Date.now();
-    await waitForActiveJobs({ redis, aggregates: batch, projectionName, kind: "map" });
-    log.write({ step: "drain-batch", tenant: tenantId, count: batch.length, durationMs: Date.now() - drainStart, kind: "map" });
-
-    // 4. CUTOFF — occurred-at bounds first, for partition pruning (see
-    //    replayBatch / getAggregateOccurredAtBounds / getBoundedCutoffs).
-    onBatchPhase("cutoff");
-    const { cutoffs, occurredAtBounds } = await getBoundedCutoffs({
-      client,
-      tenantId,
-      aggregateTypes: [...new Set(batch.map((a) => a.aggregateType))],
-      aggregateIds: batch.map((a) => a.aggregateId),
-      eventTypes: projection.definition.eventTypes,
-    });
-
-    const withCutoffKeys: string[] = [];
-    const withoutCutoffKeys: string[] = [];
-    for (const aggKey of aggKeys) {
-      if (cutoffs.has(aggKey)) {
-        withCutoffKeys.push(aggKey);
-      } else {
-        withoutCutoffKeys.push(aggKey);
-      }
-    }
-
-    log.write({ step: "cutoff-batch", tenant: tenantId, count: batch.length, withEvents: withCutoffKeys.length, kind: "map" });
-
-    if (withoutCutoffKeys.length > 0) {
-      await unmarkBatch({ redis, projectionName, aggKeys: withoutCutoffKeys });
-    }
-
-    if (withCutoffKeys.length === 0) {
-      onBatchPhase("unmark");
-      await unpauseProjection({ redis, pauseKey: projection.pauseKey });
-      return { eventsReplayed: 0 };
-    }
-
-    await markCutoffBatch({ redis, projectionName, cutoffs });
-
-    // 5. REPLAY — buffer through MapAccumulator so the WRITE phase can flush
-    //    via `store.bulkAppend` instead of awaiting one INSERT per event.
-    //    For ClickHouse-backed AppendStores this turns N round-trips into
-    //    a small number of chunked bulk inserts.
-    onBatchPhase("replay", 0);
-    const maxCutoffEventId = [...cutoffs.values()]
-      .map((c) => c.eventId)
-      .reduce((a, b) => (a > b ? a : b));
-    const aggregateIds = batch
-      .filter((a) => cutoffs.has(aggregateKey(a)))
-      .map((a) => a.aggregateId);
-
-    const accumulator = new MapAccumulator(projection.definition, this.accumulatorOpts);
-    let cursorEventId = "";
-    let eventsProcessed = 0;
-    const replayStart = Date.now();
-
-    while (true) {
-      const events = await batchLoadAggregateEvents({
-        client,
-        tenantId,
-        aggregateIds,
-        eventTypes: projection.definition.eventTypes,
-        maxCutoffEventId,
-        cursorEventId,
-        batchSize,
-        occurredAtBounds,
-      });
-
-      if (events.length === 0) break;
-
-      for (const e of events) {
-        const key = aggregateKey({
-          tenantId: e.tenantId,
-          aggregateType: e.aggregateType,
-          aggregateId: e.aggregateId,
-        });
-        const cutoff = cutoffs.get(key);
-        if (cutoff != null && isAtOrBeforeCutoff(e.timestamp, e.id, cutoff.timestamp, cutoff.eventId)) {
-          await accumulator.apply(e);
-          eventsProcessed++;
-        }
-      }
-
-      // Emit progress once per loaded page, matching the fold path's cadence
-      // — not once per event, which would hammer the progress callback.
-      onBatchPhase("replay", eventsProcessed);
-
-      const lastEvent = events[events.length - 1];
-      if (lastEvent) cursorEventId = lastEvent.id;
-      if (events.length < batchSize) break;
-    }
-
-    log.write({
-      step: "replay-batch",
-      tenant: tenantId,
-      count: withCutoffKeys.length,
-      eventsProcessed,
-      durationMs: Date.now() - replayStart,
-      kind: "map",
-    });
-
-    // 6. WRITE — flush accumulated records via bulkAppend (or sequential
-    //    append fallback, which carries each record's per-event context).
-    onBatchPhase("write", eventsProcessed);
-    await accumulator.flush();
-
-    // 7. COMPLETE + UNPAUSE — terminal `done:` markers preserve the cutoff
-    //    boundary so a job staged but never active during the pause doesn't
-    //    re-run events at/before the cutoff (double-write + double reactor
-    //    dispatch) after unpause. See the fold path for the full rationale.
-    onBatchPhase("unmark", eventsProcessed);
-    await markCompletedBatch({ redis, projectionName, cutoffs });
-    await unpauseProjection({ redis, pauseKey: projection.pauseKey });
-    log.write({ step: "unmark-batch", tenant: tenantId, count: withCutoffKeys.length, kind: "map" });
-
-    return { eventsReplayed: eventsProcessed };
-  }
-
-  async cleanup(projectionName: string): Promise<void> {
-    await cleanupAll({ redis: this.deps.redis, projectionName });
-  }
-
-  async checkPreviousRun(projectionName: string): Promise<{ completedCount: number; markerCount: number }> {
-    return hasPreviousRun({ redis: this.deps.redis, projectionName });
-  }
-
   async replayOptimized(
     config: ReplayConfig,
     callbacks?: ReplayCallbacks & { log?: ReplayLogWriter },
   ): Promise<ReplayResult> {
-    const log = callbacks?.log ?? nullLog;
-    const aggregateBatchSize = config.aggregateBatchSize ?? 1000;
-    const concurrency = config.concurrency ?? 10;
-
-    const startTime = Date.now();
-    let totalEventsReplayed = 0;
-    let totalBatchErrors = 0;
-    let firstError: string | undefined;
-    const touchedTenants = new Set<string>();
-
-    const mapProjections = config.mapProjections ?? [];
-
-    // Progress/batch reporting kind: "map" for map-only runs, otherwise
-    // "fold" (fold-only and mixed runs — fold is the dominant kind).
-    const runProjectionKind: ProjectionKind =
-      config.projections.length === 0 && mapProjections.length > 0 ? "map" : "fold";
-
-    // 1. Discover: single pass using the union of all event types (fold + map)
-    const allEventTypesForDiscovery = new Set<string>();
-    for (const p of config.projections) {
-      for (const et of p.definition.eventTypes) allEventTypesForDiscovery.add(et);
-    }
-    for (const p of mapProjections) {
-      for (const et of p.definition.eventTypes) allEventTypesForDiscovery.add(et);
-    }
-
-    const allProjectionNames = [
-      ...config.projections.map((p) => p.projectionName),
-      ...mapProjections.map((p) => p.projectionName),
-    ];
-
-    // eventTypes per projection — used to attach only the projections whose
-    // event types actually occur on each discovered aggregate. Without this,
-    // every aggregate would get cutoff/pending markers (and completion
-    // requirements) for unrelated projections that share no event types.
-    const eventTypesByProjection = new Map<string, Set<string>>();
-    for (const p of config.projections) {
-      eventTypesByProjection.set(p.projectionName, new Set(p.definition.eventTypes));
-    }
-    for (const p of mapProjections) {
-      eventTypesByProjection.set(p.projectionName, new Set(p.definition.eventTypes));
-    }
-
-    const aggregateProjectionMap = new Map<
-      string,
-      { tenantId: string; aggregateId: string; aggregateType: string; projections: string[] }
-    >();
-
-    const discoveryTargets = config.tenantIds.length > 0 ? config.tenantIds : [undefined];
-    for (const tenantId of discoveryTargets) {
-      const client = await this.resolveClient(tenantId ?? "default");
-      const aggregates = await discoverAffectedAggregates({
-        client,
-        eventTypes: [...allEventTypesForDiscovery],
-        sinceMs: new Date(config.since).getTime(),
-        tenantId,
-      });
-      for (const agg of aggregates) {
-        const key = aggregateKey(agg);
-        if (!aggregateProjectionMap.has(key)) {
-          const aggEventTypes = new Set(agg.eventTypes);
-          const matchedProjections = allProjectionNames.filter((projName) => {
-            const projEventTypes = eventTypesByProjection.get(projName)!;
-            for (const et of aggEventTypes) {
-              if (projEventTypes.has(et)) return true;
-            }
-            return false;
-          });
-          if (matchedProjections.length === 0) continue;
-          aggregateProjectionMap.set(key, {
-            tenantId: agg.tenantId,
-            aggregateId: agg.aggregateId,
-            aggregateType: agg.aggregateType,
-            projections: matchedProjections,
-          });
-        }
-      }
-    }
-
-    // Filter to specific aggregate IDs if provided (single-aggregate replay)
-    if (config.aggregateIds && config.aggregateIds.length > 0) {
-      const allowedIds = new Set(config.aggregateIds);
-      for (const [key, entry] of aggregateProjectionMap) {
-        if (!allowedIds.has(entry.aggregateId)) {
-          aggregateProjectionMap.delete(key);
-        }
-      }
-    }
-
-    if (aggregateProjectionMap.size === 0) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0 };
-    }
-
-    if (config.dryRun) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0 };
-    }
-
-    // Build lookups from projectionName to registered projection
-    const projectionByName = new Map<string, RegisteredFoldProjection>();
-    for (const p of config.projections) {
-      projectionByName.set(p.projectionName, p);
-    }
-    const mapProjectionByName = new Map<string, RegisteredMapProjection>();
-    for (const p of mapProjections) {
-      mapProjectionByName.set(p.projectionName, p);
-    }
-
-    // Get completed sets for all projections (fold + map)
-    const completedSets = new Map<string, Set<string>>();
-    for (const p of [...config.projections, ...mapProjections]) {
-      const completed = await getCompletedSet({ redis: this.deps.redis, projectionName: p.projectionName });
-      completedSets.set(p.projectionName, completed);
-    }
-
-    // Filter out aggregates completed for ALL their relevant projections
-    const allAggregateKeys = [...aggregateProjectionMap.keys()];
-    const remaining: string[] = [];
-    let skippedCount = 0;
-
-    for (const key of allAggregateKeys) {
-      const entry = aggregateProjectionMap.get(key)!;
-      const allCompleted = entry.projections.every((projName) => {
-        const completed = completedSets.get(projName);
-        return completed?.has(key) ?? false;
-      });
-      if (allCompleted) {
-        skippedCount++;
-      } else {
-        remaining.push(key);
-      }
-    }
-
-    if (remaining.length === 0) {
-      return { aggregatesReplayed: 0, totalEvents: 0, batchErrors: 0 };
-    }
-
-    // 2. Pause + drain happen PER BATCH inside the loop below (ADR-015: the
-    //    pause window is "seconds per batch", not the whole run — a full-run
-    //    pause froze live processing for as long as the replay took). The
-    //    replay marker protocol (pending/cutoff/done) keeps replayed
-    //    aggregates correct across the unpaused gaps between batches.
-    const allProjectionsToPause = [...config.projections, ...mapProjections];
-    const pausedProjectionEntries = allProjectionsToPause.map((p) => ({
-      projectionName: p.projectionName,
-      kind: p.kind,
-    }));
-
-    const runTenantCount = new Set(
-      remaining.map((key) => aggregateProjectionMap.get(key)!.tenantId),
-    ).size;
-
-    const totalBatches = Math.ceil(remaining.length / aggregateBatchSize);
-    let aggregatesCompleted = skippedCount;
-
-    for (let i = 0; i < remaining.length; i += aggregateBatchSize) {
-      const batchKeys = remaining.slice(i, i + aggregateBatchSize);
-      const batchNum = Math.floor(i / aggregateBatchSize) + 1;
-      const batchStartTime = Date.now();
-
-      const batchAggregates: DiscoveredAggregate[] = batchKeys.map((key) => {
-        const entry = aggregateProjectionMap.get(key)!;
-        return {
-          tenantId: entry.tenantId,
-          aggregateType: entry.aggregateType,
-          aggregateId: entry.aggregateId,
-        };
-      });
-
-      const progress: ReplayProgress = {
-        phase: "replaying",
-        currentProjectionName: allProjectionNames.join("+"),
-        currentProjectionKind: runProjectionKind,
-        currentProjectionIndex: 0,
-        totalProjections: allProjectionNames.length,
-        totalAggregates: allAggregateKeys.length,
-        tenantCount: runTenantCount,
-        currentBatch: batchNum,
-        totalBatches,
-        batchAggregates: batchKeys.length,
-        batchPhase: "pause",
-        batchEventsProcessed: 0,
-        aggregatesCompleted,
-        totalEventsReplayed,
-        elapsedSec: (Date.now() - startTime) / 1000,
-        skippedCount,
-        batchErrors: totalBatchErrors,
-        firstError,
-      };
-
-      const emit = () => {
-        progress.elapsedSec = (Date.now() - startTime) / 1000;
-        callbacks?.onProgress?.({ ...progress });
-      };
-
-      emit();
-
-      let batchResult: { eventsReplayed: number };
-      try {
-        // Pause only for this batch's window. The pause loop lives INSIDE the
-        // try/finally so a mid-loop pauseProjection failure still unpauses
-        // whatever was already paused (unpauseProjection is an idempotent
-        // SREM, so unpausing never-paused projections is safe).
-        for (const p of allProjectionsToPause) {
-          await pauseProjection({
-            redis: this.deps.redis,
-            pauseKey: p.pauseKey,
-          });
-        }
-        log.write({
-          step: "pause-batch",
-          batch: batchNum,
-          projections: allProjectionNames,
-        });
-
-        // Drain only THIS batch's aggregates — not every discovered aggregate.
-        progress.batchPhase = "drain";
-        emit();
-        await waitForAllActiveJobs({
-          redis: this.deps.redis,
-          aggregates: batchAggregates,
-          projections: pausedProjectionEntries,
-        });
-        log.write({ step: "drain-batch", batch: batchNum, aggregateCount: batchAggregates.length });
-
-        batchResult = await this.replayBatchOptimized({
-          batchKeys,
-          aggregateProjectionMap,
-          projectionByName,
-          mapProjectionByName,
-          concurrency,
-          log,
-          onBatchPhase: (phase, eventsProcessed) => {
-            progress.batchPhase = phase;
-            if (eventsProcessed !== undefined) {
-              progress.batchEventsProcessed = eventsProcessed;
-              progress.totalEventsReplayed = totalEventsReplayed + eventsProcessed;
-            }
-            emit();
-          },
-        });
-      } catch (error) {
-        totalBatchErrors++;
-        const errorMsg = error instanceof Error ? error.message : String(error);
-        if (!firstError) firstError = errorMsg;
-        log.write({ step: "error", batch: batchNum, error: errorMsg });
-
-        // Before the emit below — cancellation can throw from onProgress, and
-        // the failed batch's markers must be gone either way. Extra HDELs for
-        // projections an aggregate wasn't marked for are no-ops.
-        await this.clearFailedBatchMarkers({
-          projectionNames: allProjectionNames,
-          aggKeys: batchKeys,
-          log,
-        });
-
-        progress.batchErrors = totalBatchErrors;
-        progress.firstError = firstError;
-        emit();
-
-        return {
-          aggregatesReplayed: aggregatesCompleted - skippedCount,
-          totalEvents: totalEventsReplayed,
-          batchErrors: totalBatchErrors,
-          firstError,
-        };
-      } finally {
-        // Unpause after EVERY batch — including the error return above — so
-        // a failed batch can never leave live processing frozen.
-        for (const p of allProjectionsToPause) {
-          await unpauseProjection({
-            redis: this.deps.redis,
-            pauseKey: p.pauseKey,
-          }).catch(() => {});
-        }
-        log.write({ step: "unpause-batch", batch: batchNum, projections: allProjectionNames });
-      }
-
-      totalEventsReplayed += batchResult.eventsReplayed;
-      aggregatesCompleted += batchKeys.length;
-
-      for (const key of batchKeys) {
-        const entry = aggregateProjectionMap.get(key)!;
-        touchedTenants.add(entry.tenantId);
-      }
-
-      callbacks?.onBatchComplete?.({
-        projectionName: allProjectionNames.join("+"),
-        projectionKind: runProjectionKind,
-        batchNum,
-        totalBatches,
-        aggregatesInBatch: batchKeys.length,
-        eventsInBatch: batchResult.eventsReplayed,
-        durationSec: (Date.now() - batchStartTime) / 1000,
-      });
-    }
-
-    // 3. Cleanup markers for all projections
-    for (const name of allProjectionNames) {
-      await cleanupAll({ redis: this.deps.redis, projectionName: name });
-    }
-
-    // 4. Trigger OPTIMIZE TABLE on touched CH tables
-    if (totalEventsReplayed > 0 && totalBatchErrors === 0) {
-      const tables = new Set<string>();
-      for (const p of config.projections) {
-        if (p.targetTable) tables.add(p.targetTable);
-      }
-      for (const p of mapProjections) {
-        if (p.targetTable) tables.add(p.targetTable);
-      }
-
-      const tenantTargets = touchedTenants.size > 0 ? [...touchedTenants] : ["default"];
-
-      for (const tenantId of tenantTargets) {
-        try {
-          const client = await this.resolveClient(tenantId);
-          for (const table of tables) {
-            await client.command({ query: `OPTIMIZE TABLE ${table}` });
-            log.write({ step: "optimize", table, tenant: tenantId });
-          }
-        } catch {
-          // Non-fatal — merge will happen eventually
-        }
-      }
-    }
-
-    return {
-      aggregatesReplayed: aggregatesCompleted - skippedCount,
-      totalEvents: totalEventsReplayed,
-      batchErrors: totalBatchErrors,
-      firstError,
-    };
+    return replayOptimized({ ctx: this.ctx, config, callbacks });
   }
 
-  private async replayBatchOptimized({
-    batchKeys,
-    aggregateProjectionMap,
-    projectionByName,
-    mapProjectionByName,
-    concurrency,
-    log,
-    onBatchPhase,
-  }: {
-    batchKeys: string[];
-    aggregateProjectionMap: Map<
-      string,
-      { tenantId: string; aggregateId: string; aggregateType: string; projections: string[] }
-    >;
-    projectionByName: Map<string, RegisteredFoldProjection>;
-    mapProjectionByName: Map<string, RegisteredMapProjection>;
-    concurrency: number;
-    log: ReplayLogWriter;
-    onBatchPhase: (phase: BatchPhase, eventsProcessed?: number) => void;
-  }): Promise<{ eventsReplayed: number }> {
-    const redis = this.deps.redis;
-
-    // Group aggregate keys per projection — each aggregate only carries the
-    // projections whose event types occur on it, so markers must be written
-    // per (projection, matching aggregates) rather than the full cross product.
-    const aggKeysByProjection = new Map<string, string[]>();
-    for (const key of batchKeys) {
-      const entry = aggregateProjectionMap.get(key)!;
-      for (const projName of entry.projections) {
-        let list = aggKeysByProjection.get(projName);
-        if (!list) {
-          list = [];
-          aggKeysByProjection.set(projName, list);
-        }
-        list.push(key);
-      }
-    }
-    const projNames = [...aggKeysByProjection.keys()];
-
-    // 1. MARK each projection for its matching aggregates
-    onBatchPhase("mark");
-    for (const [projName, projAggKeys] of aggKeysByProjection) {
-      await markPendingBatch({ redis, projectionName: projName, aggKeys: projAggKeys });
-    }
-    log.write({ step: "mark-batch-multi", count: batchKeys.length, projections: projNames });
-
-    // 2. CUTOFF — get cutoffs per tenant, per aggregate
-    onBatchPhase("cutoff");
-    const byTenant = new Map<string, Array<{ key: string; aggregateId: string; aggregateType: string; projections: string[] }>>();
-    for (const key of batchKeys) {
-      const entry = aggregateProjectionMap.get(key)!;
-      let list = byTenant.get(entry.tenantId);
-      if (!list) {
-        list = [];
-        byTenant.set(entry.tenantId, list);
-      }
-      list.push({ key, aggregateId: entry.aggregateId, aggregateType: entry.aggregateType, projections: entry.projections });
-    }
-
-    // Collect ALL event types across all projections (fold + map) for cutoff queries
-    const allEventTypes = new Set<string>();
-    for (const projName of projNames) {
-      const foldProj = projectionByName.get(projName);
-      if (foldProj) {
-        for (const et of foldProj.definition.eventTypes) allEventTypes.add(et);
-      }
-      const mapProj = mapProjectionByName.get(projName);
-      if (mapProj) {
-        for (const et of mapProj.definition.eventTypes) allEventTypes.add(et);
-      }
-    }
-
-    // Per-tenant queries are independent — run them in parallel instead of
-    // serially awaiting one tenant at a time. Each tenant first computes its
-    // occurred-at bounds (cheap, key-column-only) so the cutoff and load
-    // queries can prune event_log's weekly partitions; see
-    // getAggregateOccurredAtBounds for the safety argument.
-    const allCutoffs = new Map<string, CutoffInfo>();
-    const boundsByTenant = new Map<string, OccurredAtBounds | undefined>();
-    await pMapLimited(
-      [...byTenant.entries()],
-      async ([tenantId, entries]) => {
-        const client = await this.resolveClient(tenantId);
-        const { cutoffs: tenantCutoffs, occurredAtBounds } = await getBoundedCutoffs({
-          client,
-          tenantId,
-          aggregateTypes: [...new Set(entries.map((e) => e.aggregateType))],
-          aggregateIds: entries.map((e) => e.aggregateId),
-          eventTypes: [...allEventTypes],
-        });
-        if (!occurredAtBounds) {
-          // Zero events for this tenant's aggregates (see getBoundedCutoffs) —
-          // no boundsByTenant entry and no allCutoffs entries, so these
-          // aggregates fall into the without-cutoff/unmark path below.
-          return;
-        }
-        boundsByTenant.set(tenantId, occurredAtBounds);
-        for (const [k, v] of tenantCutoffs) {
-          allCutoffs.set(k, v);
-        }
-      },
-      concurrency,
-    );
-
-    // Split into with/without cutoffs
-    const withCutoffKeys: string[] = [];
-    const withoutCutoffKeys: string[] = [];
-    for (const key of batchKeys) {
-      if (allCutoffs.has(key)) {
-        withCutoffKeys.push(key);
-      } else {
-        withoutCutoffKeys.push(key);
-      }
-    }
-
-    if (withoutCutoffKeys.length > 0) {
-      const withoutCutoffSet = new Set(withoutCutoffKeys);
-      for (const [projName, projAggKeys] of aggKeysByProjection) {
-        await unmarkBatch({
-          redis,
-          projectionName: projName,
-          aggKeys: projAggKeys.filter((k) => withoutCutoffSet.has(k)),
-        });
-      }
-    }
-
-    if (withCutoffKeys.length === 0) {
-      onBatchPhase("unmark");
-      return { eventsReplayed: 0 };
-    }
-
-    for (const [projName, projAggKeys] of aggKeysByProjection) {
-      const projCutoffs = new Map<string, CutoffInfo>();
-      for (const key of projAggKeys) {
-        const cutoff = allCutoffs.get(key);
-        if (cutoff) projCutoffs.set(key, cutoff);
-      }
-      await markCutoffBatch({ redis, projectionName: projName, cutoffs: projCutoffs });
-    }
-
-    // 3. REPLAY — load events per tenant, apply all relevant projections
-    onBatchPhase("replay", 0);
-
-    // Create one accumulator per projection (fold or map)
-    const foldAccumulators = new Map<string, FoldAccumulator>();
-    const mapAccumulators = new Map<string, MapAccumulator>();
-    for (const projName of projNames) {
-      const foldProj = projectionByName.get(projName);
-      if (foldProj) {
-        foldAccumulators.set(projName, new FoldAccumulator(foldProj.definition, this.accumulatorOpts));
-      }
-      const mapProj = mapProjectionByName.get(projName);
-      if (mapProj) {
-        mapAccumulators.set(projName, new MapAccumulator(mapProj.definition, this.accumulatorOpts));
-      }
-    }
-
-    // Load events grouped by tenant (one CH query per tenant, in parallel).
-    const allEvents = new Map<string, ReplayEvent[]>();
-
-    await pMapLimited(
-      [...byTenant.entries()],
-      async ([tenantId, entries]) => {
-        const aggIds = entries
-          .filter((e) => allCutoffs.has(e.key))
-          .map((e) => e.aggregateId);
-
-        if (aggIds.length === 0) return;
-
-        const client = await this.resolveClient(tenantId);
-        const tenantEvents = await loadEventsForAggregatesBulk({
-          client,
-          tenantId,
-          aggregateIds: aggIds,
-          cutoffs: allCutoffs,
-          occurredAtBounds: boundsByTenant.get(tenantId),
-        });
-
-        for (const [aggKey, events] of tenantEvents) {
-          allEvents.set(aggKey, events);
-        }
-      },
-      concurrency,
-    );
-
-    // Apply all relevant projections per aggregate — with concurrency
-    let eventsProcessed = 0;
-    let aggregatesApplied = 0;
-    const totalToApply = withCutoffKeys.length;
-
-    await pMapLimited(withCutoffKeys, async (aggKey) => {
-      const events = allEvents.get(aggKey) ?? [];
-      const entry = aggregateProjectionMap.get(aggKey)!;
-
-      for (const event of events) {
-        for (const projName of entry.projections) {
-          const foldAcc = foldAccumulators.get(projName);
-          if (foldAcc) foldAcc.apply(event);
-
-          const mapAcc = mapAccumulators.get(projName);
-          if (mapAcc) await mapAcc.apply(event);
-        }
-        eventsProcessed++;
-      }
-
-      // Throttled progress: emit every N aggregates plus the batch's last —
-      // never once per aggregate (each emit persists status to Redis).
-      aggregatesApplied++;
-      if (
-        aggregatesApplied % PROGRESS_EMIT_EVERY_AGGREGATES === 0 ||
-        aggregatesApplied === totalToApply
-      ) {
-        onBatchPhase("replay", eventsProcessed);
-      }
-    }, concurrency);
-
-    // 4. WRITE — flush all accumulators (fold states + map records in bulk)
-    onBatchPhase("write", eventsProcessed);
-    for (const [_projName, acc] of foldAccumulators) {
-      await acc.flush();
-    }
-    for (const [_projName, acc] of mapAccumulators) {
-      await acc.flush();
-    }
-
-    log.write({
-      step: "replay-batch-optimized",
-      aggregates: withCutoffKeys.length,
-      eventsProcessed,
-      projections: projNames,
-    });
-
-    // 5. COMPLETE — terminal `done:` markers per projection (not HDEL), each
-    //    preserving its aggregate's cutoff boundary so a job staged but never
-    //    active during the pause is still skipped for events at/before the
-    //    cutoff after unpause. See replayProjection for the full rationale.
-    onBatchPhase("unmark", eventsProcessed);
-    const withCutoffSet = new Set(withCutoffKeys);
-    for (const [projName, projAggKeys] of aggKeysByProjection) {
-      const projCutoffs = new Map<string, CutoffInfo>();
-      for (const key of projAggKeys) {
-        if (!withCutoffSet.has(key)) continue;
-        const cutoff = allCutoffs.get(key);
-        if (cutoff) projCutoffs.set(key, cutoff);
-      }
-      await markCompletedBatch({ redis, projectionName: projName, cutoffs: projCutoffs });
-    }
-    log.write({ step: "unmark-batch-multi", count: withCutoffKeys.length, projections: projNames });
-
-    return { eventsReplayed: eventsProcessed };
+  async cleanup(projectionName: string): Promise<void> {
+    await cleanupAll({ redis: this.ctx.redis, projectionName });
   }
-}
 
-async function pMapLimited<T>(
-  items: T[],
-  fn: (item: T) => Promise<void>,
-  concurrency: number,
-): Promise<void> {
-  const executing = new Set<Promise<void>>();
-  for (const item of items) {
-    const p = fn(item).then(() => {
-      executing.delete(p);
-    });
-    executing.add(p);
-    if (executing.size >= concurrency) await Promise.race(executing);
+  async checkPreviousRun(projectionName: string): Promise<{ completedCount: number; markerCount: number }> {
+    return hasPreviousRun({ redis: this.ctx.redis, projectionName });
   }
-  await Promise.all(executing);
 }
